+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - "Tutorial" C99La_avis added to get a quick feedback from the students.
 
+- Praise sentences in English are eliminated from result outputs of {gradethis}.
+
 ## Changes in version 2020.5.2
 
 - devel/ directory for learnrs and Shiny apps in development